Gordian Knot
I don't even want to describe the individual changes. Let's just say I did a lot of linting, moving, and fixing. Now nodes are meant to be finalised before used, however, so they're TECHNICALLY not compatible 100%. This saves a lot of fixing every operation, though - it can be done at the end in one clean sweep for performance reasons.