Add unit test to attempt to parse files
The current output is extremely problematic. However, this is supremely useful for testing: I can see exactly what code looks bad.

While I will probably implement a few practical niceties (such as the first string in the function being a """), formatting is not the main concern here: for that, just use Black et al.

So long as we can re-write a file twice and get the same result, we have preserved all meaning and thus are successful. It's... still a ways off.